chore: update dependencies
ssh-key is now re-exported by ssh-agent-lib, so no longer required as a
separate dependency.

serde shouldn't have to be declared as a dependency since it is
re-exported by clap-serde-derive, but there seems to be a bug in the
clap-serde-derive proc macro that prevents rustc from resolving the
module path to serde.

Signed-off-by: Ross Williams <ross@ross-williams.net>